requirements.txt
asgiref==3.3.1
Django==3.0.5
djangorestframework==3.12.2
djongo==1.3.3
pymongo==3.11.2
pytz==2020.5
sqlparse==0.2.4
